Much needed overhaul!
This is being done in combination with an Overseer rewrite. As far as Echo is concerned a lot of the process has been redone. Echo now must provide an integration token generated in Overseer otherwise Overseer will refuse its websocket connection. Uploads now use an ID provided by Overseer for the file names, so user's original file names are no longer used at all. This lets us just validate that we're using a UUID for the file path instead of having to check isPathInside for security.

Downloads now will only be served to clients that provide an integration token. A message is sent to Overseer asking it to validate the token gives them permission to download the file before serving it.

Uploads will only be accepted if they provide an ID Echo is expecting (which is given to Echo by Overseer before the upload starts).

Items in Overseer are not implicitly created because they exist in Echo, so when emitting a 'refresh' to overseer we now only provide the disk usage stats.

This app was originally generated using the express generator so it had a lot of dependencies that were never used. I've removed them! The important dependencies have been updated as well.

Configuration is now done through a .env file instead of a config.json file.